requirements.txt
langchain-community
langchain-core
python-dotenv
streamlit
langchain
ollama
langsmith
